feat: API

+++ api/src/app/controllers/CategoryController.js
@@ -1,4 +1,5 @@
 const CategoryRepository = require("../repositories/CategoriesRepositories");
+const isValidUUID = require("../utils/isValidUUID");
 
 class CategoryController {
   //Show all categories
@@ -22,7 +23,16 @@ class CategoryController {
   //Show a category
   async show(req, res) {
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid category ID" });
+    }
+
     const category = await CategoryRepository.findById(id);
+
+    if (!category) {
+      return res.status(404).json({ error: "Category not found!" });
+    }
     res.json(category);
   }
 
@@ -30,6 +40,10 @@ class CategoryController {
   async update(req, res) {
     const { id } = req.params;
     const { name } = req.body;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid category ID" });
+    }
 
     const categoryExists = await CategoryRepository.findById(id);
 
@@ -45,6 +59,10 @@ class CategoryController {
   //Delete a category
   async delete(req, res) {
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid cate;gory ID" });
+    }
 
     await CategoryRepository.delete(id);
